feat(cadl): initial detection of lro final state via
- sync enum values of typescript `OperationFinalStateVia` with the corresponding definition in CSharp
- add logic to parse the responses to find the `@pollingLocation` decorator and check the header value to determine whether the final state via is Location or OperationLocation
- refactor to move all operation functions into a new `lib/operation.ts`
- regenerate test cases

resolve #2983
