+++ src/CADL.Extension/Emitter.Csharp/src/type/OperationFinalStateVia.ts
@@ -4,5 +4,7 @@
 export enum OperationFinalStateVia {
     AzureAsyncOperation,
     Location,
-    OriginalUri
+    OriginalUri,
+    OperationLocation,
+    LocationOverride,
 }
